client & api: move majority of webhook stuff to unethical API, use lombok to remove a bunch of boiler plate stuff
