+++ runelite-api/src/main/java/net/unethicalite/api/discord/embeds/EmbedAuthor.java
@@ -1,0 +1,16 @@
+package net.unethicalite.api.discord.embeds;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class EmbedAuthor
+{
+    private String name;
+    private String url;
+    private String iconUrl;
+
+}
