+++ runelite-api/src/main/java/net/unethicalite/api/discord/embeds/EmbedField.java
@@ -1,0 +1,16 @@
+package net.unethicalite.api.discord.embeds;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class EmbedField
+{
+    private String name;
+    private String value;
+    private boolean inline;
+
+}
