+++ runelite-api/src/main/java/net/unethicalite/api/discord/embeds/EmbedFooter.java
@@ -1,0 +1,15 @@
+package net.unethicalite.api.discord.embeds;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class EmbedFooter
+{
+    private String text;
+    private String iconUrl;
+
+}
