--- runelite-client/src/main/java/net/runelite/client/discord/webhook/embeds/EmbedFooter.java
@@ -1,23 +1,0 @@
-package net.runelite.client.discord.webhook.embeds;
-
-public class EmbedFooter
-{
-    private String text;
-    private String iconUrl;
-
-    public EmbedFooter(String text, String iconUrl)
-    {
-        this.text = text;
-        this.iconUrl = iconUrl;
-    }
-
-    public String getText()
-    {
-        return text;
-    }
-
-    public String getIconUrl()
-    {
-        return iconUrl;
-    }
-}
