--- runelite-client/src/main/java/net/runelite/client/discord/webhook/embeds/EmbedImage.java
@@ -1,16 +1,0 @@
-package net.runelite.client.discord.webhook.embeds;
-
-public class EmbedImage
-{
-    private final String url;
-
-    public EmbedImage(String url)
-    {
-        this.url = url;
-    }
-
-    public String getUrl()
-    {
-        return url;
-    }
-}
