--- runelite-client/src/main/java/net/runelite/client/discord/webhook/embeds/EmbedThumbnail.java
@@ -1,16 +1,0 @@
-package net.runelite.client.discord.webhook.embeds;
-
-public class EmbedThumbnail
-{
-    private final String url;
-
-    public EmbedThumbnail(String url)
-    {
-        this.url = url;
-    }
-
-    public String getUrl()
-    {
-        return url;
-    }
-}
